requirements.txt
chainlit
openai
langchain
langchain_openai
langchain_core
python-dotenv
sqlalchemy
psycopg2-binary
crewai